requirements.txt
streamlit
openai
llama-index
nltk
pinecone-client
langchain
streamlit-extras